requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
defusedxml==0.5.0
Django==2.1.7
django-admin-sortable2==0.7
django-appconf==1.0.3
django-bootstrap4==0.0.7
django-braces==1.13.0
django-ckeditor==5.6.1
django-cleanup==3.1.0
django-cors-headers==2.4.0
django-crontab==0.7.1
django-extra-fields==1.0.0
django-filter==2.1.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-mptt-admin==0.6.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-select2==6.3.1
django-templated-mail==1.1.1
djangorestframework==3.9.1
easy-thumbnails==2.6
Fabric3==1.14.post1
iamport-rest-client==0.7.0
idna==2.8
oauthlib==3.0.1
paramiko==2.4.2
paypalrestsdk==1.13.1
Pillow==6.0.0
psycopg2-binary==2.7.7
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
sqlparse==0.3.0
urllib3==1.24.1
Wand==0.5.2
